Scripts/requirements.txt
certifi==2019.9.11
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
mysql-connector==2.2.9
numpy==1.17.2
openpyxl==2.6.3
pandas==0.25.1
pdfrw==0.4
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.5
xlrd==1.2.0
